backend/requirement.txt
Django>=4.2,<5.0
psycopg2-binary
python-dotenv

# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
chromadb==0.4.15
sentence-transformers==2.2.2
spacy==3.7.2
langdetect==1.0.9
replicate==0.22.0
beautifulsoup4==4.12.2
numpy==1.24.3
urllib3==1.26.16
python-dotenv==1.0.0
corsheaders==4.3.1